Try to fix GUI button sometimes overflowing
The issue only happens sometimes, and I didn't see it when trying to fix
it. If I see the problem again I'll give it another shot.

Setting layout to "List" for gui_container has nothing to do with it (I
think). It seems like this was the default value made explicit. The "Flow"/"Fill" just made the buttons not show up at all (off screen?).